Refactor random_generator_base
