+++ onnxruntime/test/common/tensor_op_test_utils.cc
@@ -8,9 +8,7 @@ namespace onnxruntime {
 namespace test {
 
 RandomValueGenerator::RandomValueGenerator(optional<RandomSeedType> seed)
-    : random_seed_{
-          seed.has_value() ? *seed : static_cast<RandomSeedType>(GetTestRandomSeed())},
-      generator_{random_seed_},
+    : RandomValueGeneratorBase(seed),
       output_trace_{__FILE__, __LINE__, "ORT test random seed: " + std::to_string(random_seed_)} {
 }
